% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello(message = "")
}
\arguments{
\item{message}{texto para impressao. default is empty.}
}
\value{
nada
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\examples{
hello()
hello("Hugo")
\dontrun{
  ler_varios_csv("caminho-ate-pasta")
}
}
